+++ code/ryzom/server/src/pd_support_service/hourly_commands.cpp
@@ -46,7 +46,7 @@ public:
 			}
 			return;
 		}
-		
+
 		// get the start time
 		time_t startTime;
 		time( &startTime );
@@ -73,7 +73,7 @@ public:
 		// execute hourly tasks
 		NLMISC::CConfigFile::CVar *commandsVar = NLNET::IService::getInstance()->ConfigFile.getVarPtr("HourlyCommands");
 		WARN_IF(commandsVar  == NULL,"'HourlyCommands' not found in cfg file");
-			
+
 		// if we have hourly commands...
 		if (commandsVar!=NULL)
 		{
@@ -104,7 +104,7 @@ public:
 		FILE* fileHandle= fopen(HourlyActivityLogFileName,"ab");
 		nlassert(fileHandle!=NULL);
 		fprintf(fileHandle,"%02u/%02u/%u CHourlyTaskScheduler: Started: %02u:%02u, Finished: %02u:%02u, Executed %u commands Started %u Jobs\n",
-			ptm->tm_mday, ptm->tm_mon+1, ptm->tm_year+1900, startTime/3600%24, startTime/60%60, endTime/3600%24, endTime/60%60, commandsVar==NULL?0:commandsVar->size(), jobsRemaining );
+			ptm->tm_mday, ptm->tm_mon+1, ptm->tm_year+1900, (uint)startTime/3600%24, (uint)startTime/60%60, (uint)endTime/3600%24, (uint)endTime/60%60, commandsVar==NULL?0:commandsVar->size(), jobsRemaining );
 		nlinfo("JobManager state: %s",CJobManager::getInstance()->getStatus().c_str());
 		fclose(fileHandle);
 	}
